Rework check-universe test
